+++ src/datatypes/qbsproducttype.ts
@@ -1,0 +1,5 @@
+export enum QbsProductType {
+    Application = 'application',
+    DynamicLibrary = 'dynamiclibrary',
+    StaticLibrary = 'staticlibrary',
+}